Refactor: rearrange components and styles, update the project structure
