+++ package.json
@@ -13,6 +13,7 @@
     "type": "git",
     "url": "https://github.com/oneteme/jquery-charts.git"
   },
+
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
