+++ src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './components/home/home.component';
+import { HomeComponent } from './layout/home/home.component';
 import { DocumentationComponent } from './features/documentation/documentation.component';
-import { ChartViewComponent } from './components/chart-view/chart-view.component';
+import { ChartViewComponent } from './features/charts/components/chart-view/chart-view.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/home', pathMatch: 'full' },
