+++ src/app/features/charts/components/chart-group/chart-group.component.scss
