+++ src/styles.scss
@@ -1,5 +1,3 @@
-/* You can add global styles to this file, and also import other style files */
-
 // Reset SCSS
 *,
 *::before,
@@ -18,12 +16,10 @@ html, body {
   color: #333;
 }
 
-// Supprime les styles par défaut des listes
 ul, ol {
   list-style: none;
 }
 
-// Suppression du style des liens
 a {
   text-decoration: none;
   color: inherit;
@@ -34,25 +30,21 @@ a {
   }
 }
 
-// Réinitialisation des boutons
 button {
   all: unset;
   cursor: pointer;
 }
 
-// Images responsives
 img {
   max-width: 100%;
   height: auto;
 }
 
-// Supprime les bordures par défaut sur les éléments de formulaire
 input, textarea, select, button {
   border: none;
   outline: none;
 }
 
-// Ajoute un style de focus accessible
 :focus {
   outline: 2px solid #007bff;
   outline-offset: 2px;
